Доработка по замечаниям наставника
