+++ js/gallery.js
@@ -7,35 +7,37 @@ const pictureContainer = document.querySelector('.pictures');
 const pictureTemplate = document.querySelector('#picture');
 
 /**
- * Отрисовывает миниатюры
- * @param {Array<Picture>} data - массив с данными для картинок
+ * Создает разметку для одной миниатюры
+ * @param {Picture} picture - данные одной картинки
+ * @returns HTMLAnchorElement - разметка одной миниатюры
  */
-const renderThumbnails = (data) => {
-  const fragment = document.createDocumentFragment();
+const createThumbnail = (picture) => {
+  const newPicture = /** @type {HTMLAnchorElement} */ (
+    pictureTemplate.content.cloneNode(true)
+  );
+  const {url, description, comments, likes} = picture;
 
-  for (const picture of data) {
+  newPicture.querySelector('.picture__img').setAttribute('src',url);
+  newPicture.querySelector('.picture__img').setAttribute('alt',description);
+  newPicture.querySelector('.picture__comments').textContent = String(comments.length);
+  newPicture.querySelector('.picture__likes').textContent = String(likes);
 
-    const newPicture = /** @type {HTMLAnchorElement} */ (
-      pictureTemplate.content.cloneNode(true)
-    );
-    const {url, description, comments, likes, id} = picture;
+  newPicture.querySelector('.picture').addEventListener('click', (event) => {
+    event.preventDefault();
+    //openModal(pictureWindow);
+    renderPopup(picture);
+  });
 
-    newPicture.querySelector('.picture__img').setAttribute('src',url);
-    newPicture.querySelector('.picture__img').setAttribute('alt',description);
-    newPicture.querySelector('.picture__comments').textContent = String(comments.length);
-    newPicture.querySelector('.picture__likes').textContent = String(likes);
-    newPicture.querySelector('.picture__img').setAttribute('id', String(id));
-
-    newPicture.querySelector('.picture').addEventListener('click', (event) => {
-      event.preventDefault();
-      renderPopup(picture);
-    });
-
-
-    fragment.appendChild(newPicture);
-  }
-  pictureContainer.appendChild(fragment);
-
+  return newPicture;
 };
 
-export {renderThumbnails};
+/**
+ * Отрисовывает миниатюры
+ * @param {Array<Picture>} data  - массив с данными для картинок
+ */
+const renderGallery = (data) => {
+  const newThumbnails = data.map(createThumbnail);
+  pictureContainer.append(...newThumbnails);
+};
+
+export {renderGallery};
